+++ spec/types/hash_spec.rb
@@ -9,6 +9,43 @@ describe Attributor::Hash do
   its(:key_type) { should be(Attributor::Object) }
   its(:value_type) { should be(Attributor::Object) }
 
+context 'attributes' do
+    context 'with an exception from the definition block' do
+      subject(:broken_model) do
+        Class.new(Attributor::Model) do
+          attributes do
+            raise 'sorry :('
+            attribute :name, String
+          end
+        end
+      end
+
+      it 'throws original exception upon first run' do
+        lambda {
+          broken_model.attributes
+        }.should raise_error(RuntimeError, 'sorry :(')
+      end
+
+      it 'throws InvalidDefinition for subsequent access' do
+        broken_model.attributes rescue nil
+
+        lambda { 
+          broken_model.attributes
+        }.should raise_error(Attributor::InvalidDefinition)
+      end
+
+      it 'throws for any attempts at using of an instance of it' do
+        broken_model.attributes rescue nil
+
+        instance = broken_model.new
+        lambda {
+          instance.name
+        }.should raise_error(Attributor::InvalidDefinition)
+      end
+
+    end
+  end
+  
   context 'default options' do
     subject(:options) { type.options }
     it 'has allow_extra false' do